+++ tools/build-package.sh
@@ -78,6 +78,7 @@ find $source_roots -type f \
   ! -path '*/.build-m10/*' \
   ! -path '*/.build-m11/*' \
   ! -path '*/.build-m12/*' \
+  ! -path '*/.build-m13/*' \
   ! -path '*/.libs/*' \
   ! -path '*/.deps/*' \
   -print | LC_ALL=C sort > "$manifest"